requirements.txt
funcsigs==1.0.2
mock==2.0.0
nltk==3.2.1
pbr==1.10.0
python-slugify==1.2.0
PyYAML==3.11
six==1.10.0
Unidecode==0.4.19
